+++ CHANGELOG.rst
@@ -45,6 +45,7 @@ Changed:
 * Generalized ascat, delly, and manta result in workflow. #708
 * Generalized workflow to eliminate duplicate entries and code. #708
 * Split Tumor-Normal ``cnvkit batch`` into individual commands
+* Moved params that are used in multiple rules to constants #711
 * Changed the way conda and non-conda bioinfo tools version are parsed